Convert Event from an enum to a trait
This opens up uses of enum so that others outside the crate can add
their own event types. Ideally we would support all the relevant events
within the crate but that's not always feasible. Having this be a trait
which just sets some fields within the group will mean that users of the
crate have the ability to implement something even if we haven't added
support explicitly.

In a further change to the API, instead of having a Builder::kind method
which takes an event. I have instead changed Builder::new to directly
take the event. This way there's never any confusion over what the event
type is if it is not specified.