Make Sample deref to Counter instead of using a macro
The macro that copied a bunch of methods across both Counter and Sampler
was rather confusing to work with. By instead having Sample dereference
to a Counter both the docs and the implementation become much easier to
read. This should have no effect on users of the library.

This is against the rust API guidelines. However, I think the benefits
here are worth it and none of the usual downsides apply since we control
both Sampler and Counter.