Add a helper type for reading and caching dynamic PMU types  (#28)
As new events get added there will be more dynamic PMU events being
added. Each of these needs to do the same thing to read the PMU type.
It is better to wrap that all up in a helper type so that people writing
PRs don't have to copy-paste the same code again and again.

This commit also converts kprobe and uprobe to use the new helper.