Move Msr to new events::x86::intel module
The msr kernel perf module is x86-specific. It is better to indicate
that to users of the crate directly. In addition, this way when more
event types come we don't have to worry about, for example, the x86 MSR
event and the ARM MSR event having a name clash.