Fix a panic within ByteBuffer::truncate
There was a reversed comparison when determining whether we should
advance past the split within the buffer. It was checking whether
a.len() <= new_len but it should have been checking whether
new_len <= a.len(). This was leading to panics whenever a sampler record
ended up being split.

I've fixed the if statement and added some test cases to cover both
split cases with a truncate.