Update all manpage links to point to mankier.com
man7.org is great but hasn't been updated in a few years. Most manpages
don't really change that often so this usually doesn't matter that much
but the perf_event_open manpage changes fairly frequently. This means
that the manpage at man7.org is missing a few options and will continue
to get worse as time goes on.

mankier.com seems to be reliably updated to the most recent kernel
version and has been going for about a decade now so it should be
reliable enough to link to from the documentation.